requirements.txt
python-dotenv==1.2.1
langgraph==1.0.3
openai==1.109.1
llama-parse==0.6.88
langchain-text-splitters==0.3.11
torch==2.9.0
transformers==4.57.1
modelscope==1.33.0
pymilvus==2.6.5